src/requirements.txt
opencv-contrib-python-headless
pillow
opencv-python

